#ROCKETMQ-281#  add file lock  for preventing repeat start mq
